+++ torch_geometric/nn/models/dimenet.py
@@ -338,9 +338,14 @@ class OutputBlock(torch.nn.Module):
         out_channels: int,
         num_layers: int,
         act: Callable,
+        output_initializer: str = 'zeros',
     ):
+        assert output_initializer in {'zeros', 'glorot_orthogonal'}
+
         super().__init__()
+
         self.act = act
+        self.output_initializer = output_initializer
 
         self.lin_rbf = Linear(num_radial, hidden_channels, bias=False)
         self.lins = torch.nn.ModuleList()
@@ -355,7 +360,10 @@ class OutputBlock(torch.nn.Module):
         for lin in self.lins:
             glorot_orthogonal(lin.weight, scale=2.0)
             lin.bias.data.fill_(0)
-        glorot_orthogonal(self.lin.weight, scale=2.0)
+        if self.output_initializer == 'zeros':
+            self.lin.weight.data.fill_(0)
+        elif self.output_initializer == 'glorot_orthogonal':
+            glorot_orthogonal(self.lin.weight, scale=2.0)
 
     def forward(self, x: Tensor, rbf: Tensor, i: Tensor,
                 num_nodes: Optional[int] = None) -> Tensor:
@@ -375,9 +383,14 @@ class OutputPPBlock(torch.nn.Module):
         out_channels: int,
         num_layers: int,
         act: Callable,
+        output_initializer: str = 'zeros',
     ):
+        assert output_initializer in {'zeros', 'glorot_orthogonal'}
+
         super().__init__()
+
         self.act = act
+        self.output_initializer = output_initializer
 
         self.lin_rbf = Linear(num_radial, hidden_channels, bias=False)
 
@@ -396,7 +409,10 @@ class OutputPPBlock(torch.nn.Module):
         for lin in self.lins:
             glorot_orthogonal(lin.weight, scale=2.0)
             lin.bias.data.fill_(0)
-        glorot_orthogonal(self.lin.weight, scale=2.0)
+        if self.output_initializer == 'zeros':
+            self.lin.weight.data.fill_(0)
+        elif self.output_initializer == 'glorot_orthogonal':
+            glorot_orthogonal(self.lin.weight, scale=2.0)
 
     def forward(self, x: Tensor, rbf: Tensor, i: Tensor,
                 num_nodes: Optional[int] = None) -> Tensor:
@@ -470,6 +486,9 @@ class DimeNet(torch.nn.Module):
             output blocks. (default: :obj:`3`)
         act (str or Callable, optional): The activation function.
             (default: :obj:`"swish"`)
+        output_initializer (str, optional): The initialization method for the
+            output layer (:obj:`"zeros"`, :obj:`"glorot_orthogonal"`).
+            (default: :obj:`"zeros"`)
     """
 
     url = ('https://github.com/klicperajo/dimenet/raw/master/pretrained/'
@@ -490,6 +509,7 @@ class DimeNet(torch.nn.Module):
         num_after_skip: int = 2,
         num_output_layers: int = 3,
         act: Union[str, Callable] = 'swish',
+        output_initializer: str = 'zeros',
     ):
         super().__init__()
 
@@ -509,8 +529,14 @@ class DimeNet(torch.nn.Module):
         self.emb = EmbeddingBlock(num_radial, hidden_channels, act)
 
         self.output_blocks = torch.nn.ModuleList([
-            OutputBlock(num_radial, hidden_channels, out_channels,
-                        num_output_layers, act) for _ in range(num_blocks + 1)
+            OutputBlock(
+                num_radial,
+                hidden_channels,
+                out_channels,
+                num_output_layers,
+                act,
+                output_initializer,
+            ) for _ in range(num_blocks + 1)
         ])
 
         self.interaction_blocks = torch.nn.ModuleList([
@@ -721,6 +747,9 @@ class DimeNetPlusPlus(DimeNet):
             output blocks. (default: :obj:`3`)
         act: (str or Callable, optional): The activation funtion.
             (default: :obj:`"swish"`)
+        output_initializer (str, optional): The initialization method for the
+            output layer (:obj:`"zeros"`, :obj:`"glorot_orthogonal"`).
+            (default: :obj:`"zeros"`)
     """
 
     url = ('https://raw.githubusercontent.com/gasteigerjo/dimenet/'
@@ -743,6 +772,7 @@ class DimeNetPlusPlus(DimeNet):
         num_after_skip: int = 2,
         num_output_layers: int = 3,
         act: Union[str, Callable] = 'swish',
+        output_initializer: str = 'zeros',
     ):
         act = activation_resolver(act)
 
@@ -760,6 +790,7 @@ class DimeNetPlusPlus(DimeNet):
             num_after_skip=num_after_skip,
             num_output_layers=num_output_layers,
             act=act,
+            output_initializer=output_initializer,
         )
 
         # We are re-using the RBF, SBF and embedding layers of `DimeNet` and
@@ -775,6 +806,7 @@ class DimeNetPlusPlus(DimeNet):
                 out_channels,
                 num_output_layers,
                 act,
+                output_initializer,
             ) for _ in range(num_blocks + 1)
         ])